Fix CircleCI hadolint artifact path
The CircleCI orb `circleci/docker` includes `hadolint`, a linter for
Dockerfiles. By default, it tries to upload all files in the path
`~/project` as artifacts after running the linter.

However, CircleCI also put all project files from the repository in this
same path, which causes the hadolint workflow to upload source files as
artifacts to CircleCI. There's an open issue for this behavior at the
CircleCI GitHub repository
(https://github.com/CircleCI-Public/docker-orb/issues/17).

This behavior was noticed after using the file `.circleci/config.yml`
from kotlin-spring-sample into the repository Creditas/archimedes (Ruby
project).

As suggested in the issue mentioned above, setting the parameter
`artifacts-path` to another path can change this unwanted behavior of
uploading source files as artifacts. I tested this parameter with the
Creditas/archimedes repository and it worked just fine. You can compare
the workflows with and without the parameter to notice the behavior on
CircleCI:

* without `artifacts-path`:
  https://app.circleci.com/jobs/github/Creditas/archimedes/4170

* with `artifacts-path`:
  https://app.circleci.com/jobs/github/Creditas/archimedes/4172

Even though archimedes isn't a fork from the kotlin-spring-sample
repository, I believe this behavior can be reproduced with a Kotlin
project.

With this commit, I propose setting the `artifacts-path` parameter in
the kotlin-spring-sample to a path which will not include any source
files.